requirements.txt
absl-py==2.1.0
alembic==1.13.1
astunparse==1.6.3
attrs==23.2.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
colorlog==6.8.2
contourpy==1.2.1
cycler==0.12.1
flatbuffers==24.3.25
fonttools==4.51.0
gast==0.4.0
google-auth==2.29.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.64.0
h5py==3.11.0
idna==3.7
importlib-metadata==7.1.0
importlib-resources==6.4.0
jax==0.4.28
jaxlib==0.4.28
joblib==1.4.2
keras==3.3.3
kiwisolver==1.4.5
markdown==3.6
markdown-it-py==3.0.0
markupsafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
mediapipe==0.10.14
ml-dtypes==0.3.2
namex==0.0.8
numpy==1.23.5
# numpy==1.26.4
oauthlib==3.2.2
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
optuna==3.6.1
packaging==24.0
pandas==2.2.2
protobuf==4.25.3
pyasn1==0.6.0
pyasn1-modules==0.4.0
pycparser==2.22
pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
pyyaml==6.0.1
requests==2.32.1
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
scikeras==0.13.0
scikit-learn==1.5.0
scipy==1.13.1
seaborn==0.13.2
six==1.16.0
sounddevice==0.4.6
sqlalchemy==2.0.30
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.5.0
tqdm==4.66.4
typing-extensions==4.11.0
urllib3==2.2.1
werkzeug==3.0.3
wrapt==1.14.1
zipp==3.18.2
# App and chatbot
streamlit>=1.35.0
langchain>=0.2.0
langchain-community>=0.2.0
openai>=0.27.0
python-dotenv>=1.0.0
dataclasses-json>=0.6.0
